+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Require `fun_with_flags 0.11`.
 * Add support for clearing boolean gates, and display when a boolan gate is missing.
+* Add support for Percentage gates.
 
 ## v0.3.0